test(log): test block index recovery
* extend recovery integration tests
* fix log message length type (short > int)
* fix recovery with existing snapshot
* configurable write buffer size and block index size
* improve and externalize writing index buffer to output stream

related to #15

+++ logstreams/src/main/java/org/camunda/tngp/logstreams/BufferedLogStreamReader.java
@@ -1,8 +1,19 @@
 package org.camunda.tngp.logstreams;
 
-import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.*;
-import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.*;
-import static org.camunda.tngp.logstreams.spi.LogStorage.*;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.HEADER_LENGTH;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.alignedLength;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.lengthOffset;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.messageOffset;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.streamIdOffset;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.typeOffset;
+import static org.camunda.tngp.dispatcher.impl.log.DataFrameDescriptor.versionOffset;
+import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.HEADER_BLOCK_LENGHT;
+import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.headerLength;
+import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.keyLengthOffset;
+import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.keyOffset;
+import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.positionOffset;
+import static org.camunda.tngp.logstreams.impl.LogEntryDescriptor.valueOffset;
+import static org.camunda.tngp.logstreams.spi.LogStorage.OP_RESULT_INSUFFICIENT_BUFFER_CAPACITY;
 
 import java.nio.ByteBuffer;
 import java.util.NoSuchElementException;
@@ -379,7 +390,7 @@ public class BufferedLogStreamReader implements LogStreamReader
         @Override
         public int getMessageLength()
         {
-            return buffer.getShort(lengthOffset(fragmentOffset));
+            return buffer.getInt(lengthOffset(fragmentOffset));
         }
 
         @Override
